fixed compile error

+++ core/src/test/java/org/fao/geonet/kernel/mef/MEFLibIntegrationTest.java
@@ -1,16 +1,6 @@
 package org.fao.geonet.kernel.mef;
 
-import static junit.framework.Assert.assertNotNull;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.fail;
-
-import java.io.File;
-import java.io.InputStream;
-import java.util.ArrayList;
-import java.util.List;
-
 import jeeves.server.context.ServiceContext;
-
 import org.apache.commons.io.FileUtils;
 import org.fao.geonet.AbstractCoreIntegrationTest;
 import org.fao.geonet.constants.Params;
@@ -21,6 +11,15 @@ import org.jdom.Element;
 import org.junit.Ignore;
 import org.junit.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+
+import java.io.File;
+import java.io.InputStream;
+import java.util.ArrayList;
+import java.util.List;
+
+import static junit.framework.Assert.assertNotNull;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.fail;
 
 /**
  * Test MEF.
@@ -45,7 +44,7 @@ public class MEFLibIntegrationTest extends AbstractCoreIntegrationTest {
         final Metadata metadata = _metadataRepo.findOne(metadataIds.get(0));
 
         assertNotNull(metadata);
-        assertEquals(admin.getId(), metadata.getSourceInfo().getOwner());
+        assertEquals(admin.getId(), metadata.getSourceInfo().getOwner().intValue());
     }
 
     @Test
@@ -66,7 +65,7 @@ public class MEFLibIntegrationTest extends AbstractCoreIntegrationTest {
             final Metadata metadata = _metadataRepo.findOne(metadataId);
 
             assertNotNull(metadata);
-            assertEquals(admin.getId(), metadata.getSourceInfo().getOwner());
+            assertEquals(admin.getId(), metadata.getSourceInfo().getOwner().intValue());
         }
     }
